refactor: 파일 구조 정리 및 import 경로 업데이트 #37
