+++ fe/src/components/fixed-costs/FixedCostsPage.tsx
@@ -8,7 +8,10 @@ import { Badge } from '@/components/ui/badge';
 import ResponsivePanel from '../panel/ResponsivePanel';
 import FixedCostSubmitForm from './FixedCostSubmitForm';
 import { IFixedRule } from '@/types/fixed-costs';
-import { fetchFixedRules, setFixedRuleActive } from '@/services/fixed-costs';
+import {
+  fetchFixedRules,
+  setFixedRuleActive,
+} from '@/services/fixed-costs/fixed-costs';
 import { toast } from 'sonner';
 import { mapFixedRuleToFormData } from '@/utils/fixed-costs';
