+++ fe/src/services/fixed-costs/fixed-costs.ts
